fix(code-templates): BigDecimal 을 값 객체로 내보내지 않는다
인사관리 모델로 생성했더니 없는 클래스를 import 했다.

    import com.poscodx.sample.leaveManagement.store.domain.vo.BigDecimal;

템플릿의 `isPrimitive` 헬퍼는 String·Integer·Long·Double·Float·Boolean·
Date 만 원시로 본다. 거기 없는 이름은 값 객체로 취급돼 생성 패키지에서
찾는다. 소문자 원시형과 UUID 가 같은 자리에서 샜고, `BigDecimal` 이
남아 있었다 — 휴가 일수처럼 소수 자리를 가진 속성에서 나온다.

UUID 와 같은 이유로 `Double` 로 내보낸다. 정밀도는 잃지만 컴파일되지
않는 코드보다는 낫다.

실측 — 252개 생성, `verify-code-templates.mjs` 11종 전부 통과
(종전 "JDK 타입을 생성 패키지에서 import 하지 않는다" 실패).

+++ api/features/code_templates/context.py
@@ -35,7 +35,11 @@ _JAVA_TYPES = {
     "uuid": "String",
     "date": "Date", "datetime": "Date", "localdate": "Date",
     "localdatetime": "Date", "timestamp": "Date",
-    "bigdecimal": "BigDecimal", "decimal": "BigDecimal", "number": "Double",
+    # BigDecimal 도 UUID 와 같은 이유로 내보낸다 — 자바 표준 타입이지만
+    # 템플릿의 원시 목록에 없어서, 그대로 두면 값 객체로 취급돼
+    # `import ….store.domain.vo.BigDecimal;` 이라는 없는 클래스를 부른다.
+    # 정밀도는 잃지만 컴파일되지 않는 코드보다는 낫다.
+    "bigdecimal": "Double", "decimal": "Double", "number": "Double",
 }
